Break data wrapper into a module and further parameterize normalization
