+++ inferelator/utils/__init__.py
@@ -1,14 +1,23 @@
-from inferelator.utils.validator import Validator, is_string
-from inferelator.utils.debug import Debug, slurm_envs, inferelator_verbose_level
-from inferelator.utils.loader import InferelatorDataLoader, DEFAULT_PANDAS_TSV_SETTINGS
-from inferelator.utils.data import (
-    InferelatorData,
+from .validator import (
+    Validator,
+    is_string
+)
+from .debug import (
+    Debug,
+    slurm_envs,
+    inferelator_verbose_level
+)
+from .loader import (
+    InferelatorDataLoader,
+    DEFAULT_PANDAS_TSV_SETTINGS
+)
+from .data import (
     df_from_tsv,
     array_set_diag,
     df_set_diag,
     melt_and_reindex_dataframe,
     make_array_2d,
-    scale_vector,
     DotProduct,
     safe_apply_to_array
 )
+from .inferelator_data import InferelatorData
